+++ fabric/src/main/java/com/evandev/reliable_requiem/ReliableRequiem.java
@@ -3,6 +3,7 @@ package com.evandev.reliable_requiem;
 import com.evandev.reliable_requiem.api.IPlayerKeptItems;
 import com.evandev.reliable_requiem.effect.MementoMoriEffect;
 import com.evandev.reliable_requiem.modules.ModEffects;
+import com.evandev.reliable_requiem.platform.Services;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.entity.event.v1.ServerPlayerEvents;
 import net.minecraft.core.Registry;
@@ -10,6 +11,7 @@ import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.ItemStack;
 
+import net.minecraft.nbt.CompoundTag;
 import java.util.Map;
 
 public class ReliableRequiem implements ModInitializer {
@@ -30,6 +32,9 @@ public class ReliableRequiem implements ModInitializer {
             CommonClass.onPlayerClone(oldPlayer, newPlayer, wasDeath);
 
             if (wasDeath) {
+                CompoundTag keptAcc = ((IPlayerKeptItems) oldPlayer).reliableRequiem$getKeptAccessories();
+                ((IPlayerKeptItems) newPlayer).reliableRequiem$setKeptAccessories(keptAcc);
+
                 Map<Integer, ItemStack> keptItems = ((IPlayerKeptItems) oldPlayer).reliableRequiem$getKeptItems();
 
                 for (Map.Entry<Integer, ItemStack> entry : keptItems.entrySet()) {
@@ -37,7 +42,13 @@ public class ReliableRequiem implements ModInitializer {
                     ItemStack stack = entry.getValue();
                     newPlayer.getInventory().setItem(slot, stack);
                 }
+
+                Services.PLATFORM.restoreKeptAccessories(newPlayer);
             }
+        });
+
+        ServerPlayerEvents.AFTER_RESPAWN.register((oldPlayer, newPlayer, alive) -> {
+            CommonClass.onPlayerRespawn(newPlayer);
         });
     }
 }