+++ forge/src/main/java/com/evandev/reliable_requiem/ReliableRequiem.java
@@ -4,12 +4,15 @@ import com.evandev.reliable_requiem.api.IPlayerKeptItems;
 import com.evandev.reliable_requiem.client.ClientConfigSetup;
 import com.evandev.reliable_requiem.effect.MementoMoriEffect;
 import com.evandev.reliable_requiem.modules.ModEffects;
+import com.evandev.reliable_requiem.platform.Services;
 import net.minecraft.core.registries.Registries;
+import net.minecraft.nbt.CompoundTag;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.item.ItemStack;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.entity.player.PlayerEvent;
+import net.minecraftforge.eventbus.api.EventPriority;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.ModLoadingContext;
@@ -51,13 +54,17 @@ public class ReliableRequiem {
         }
     }
 
-    @SubscribeEvent
+    // curios doesn't like working unless we use a low priority here
+    @SubscribeEvent(priority = EventPriority.LOWEST)
     public void onPlayerClone(PlayerEvent.Clone event) {
         if (event.getEntity() instanceof ServerPlayer newPlayer && event.getOriginal() instanceof ServerPlayer oldPlayer) {
 
             CommonClass.onPlayerClone(oldPlayer, newPlayer, event.isWasDeath());
 
             if (event.isWasDeath()) {
+                CompoundTag keptAcc = ((IPlayerKeptItems) oldPlayer).reliableRequiem$getKeptAccessories();
+                ((IPlayerKeptItems) newPlayer).reliableRequiem$setKeptAccessories(keptAcc);
+
                 Map<Integer, ItemStack> keptItems = ((IPlayerKeptItems) oldPlayer).reliableRequiem$getKeptItems();
 
                 for (Map.Entry<Integer, ItemStack> entry : keptItems.entrySet()) {
@@ -65,7 +72,16 @@ public class ReliableRequiem {
                     ItemStack stack = entry.getValue();
                     newPlayer.getInventory().setItem(slot, stack);
                 }
+
+                Services.PLATFORM.restoreKeptAccessories(newPlayer);
             }
+        }
+    }
+
+    @SubscribeEvent
+    public void onPlayerRespawn(PlayerEvent.PlayerRespawnEvent event) {
+        if (event.getEntity() instanceof ServerPlayer player) {
+            CommonClass.onPlayerRespawn(player);
         }
     }
 }