+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2026-07-07
+
+### Added
+
+- Added support for keeping Curios/Trinkets on death.
+
+### Fixed
+
+- Fixed client/server desync with Memento Mori.
+
 ## [1.0.3] - 2026-05-26
 
 ### Fixed